# Docs/kindohub-api-spec.yaml
openapi: 3.0.3
info:
  title: KindoHub API
  version: 1.0.0
  description: API REST para la gestion integral de una asociacion de padres y madres (APA/AMPA). Incluye gestion de familias, alumnos, cursos, anotaciones, usuarios, formas de pago, estados de asociado y sistema de auditoria con logs.
  contact:
    name: KindoHub Development Team
    email: dev@kindohub.com
  license:
    name: Proprietary
    url: https://kindohub.com/license
servers:
  - url: https://api.kindohub.com
    description: Servidor de produccion
  - url: https://staging-api.kindohub.com
    description: Servidor de staging
  - url: http://localhost:5000
    description: Desarrollo local
...
paths:
...
  /api/familias/{id}:
    get:
      tags:
        - Familias
      summary: Obtener familia por ID
      operationId: getFamilyById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Familia encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamiliaDto'
        '404':
          description: Familia no encontrada
  /api/familias:
    get:
      tags:
        - Familias
      summary: Listar todas las familias
      operationId: getAllFamilies
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de familias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FamiliaDto'
...
  /api/log:
    get:
      tags:
        - Logs
      summary: Obtener todos los logs
      operationId: getAllLogs
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LogDto'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FamiliaDto:
      type: object
      properties:
        Id:
          type: integer
        Nombre:
          type: string
        Email:
          type: string
          format: email
        Telefono:
          type: string
        Direccion:
          type: string
        IdFormaPago:
          type: integer
...
    LogDto:
      type: object
      properties:
        Id:
          type: integer
        Message:
          type: string
        Level:
          type: string
        TimeStamp:
          type: string
          format: date-time
        Username:
          type: string
security:
  - BearerAuth: []